backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.8.0.post1",
    "fastapi[standard]>=0.115.7",
    "mistralai>=1.5.0",
    "numpy>=1.23.5",
    "opencv-python>=4.8.1.78",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "pydantic-core>=2.27.2",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "ruff>=0.9.3",
    "tensorflow==2.12.0",
    "uvicorn>=0.34.0",
]
